+++ src/test/java/com/google/error_analyzer/backend/IndexNameTest.java
@@ -1,11 +1,8 @@
 /**Copyright 2019 Google LLC
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
     https://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,12 +12,9 @@ limitations under the License.*/
 package com.google.error_analyzer;
 
 import com.google.error_analyzer.backend.IndexName;
-import com.google.error_analyzer.backend.LogDao;
-import com.google.error_analyzer.backend.MockLogDao;
-import com.google.error_analyzer.backend.StoreLogs;
 import java.io.IOException;
-import java.util.ArrayList;
 import javax.servlet.http.*;
+import org.elasticsearch.common.text.Text;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
@@ -33,45 +27,43 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.runners.MockitoJUnitRunner;
-import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-
-/*this class contains tests for the methods used for 
-storing logs to the database*/
-public final class StoreLogTest {
-    private StoreLogs storeLogs;
+public class IndexNameTest {
     private Cookie cookie;
 
     @Mock
     HttpServletRequest request;
 
+    @Rule
+    public MockitoRule rule = MockitoJUnit.rule();
+
     @Before
     public void setUp() {
-        storeLogs = new StoreLogs();
-        storeLogs.logDao = new MockLogDao();
         request = Mockito.mock(HttpServletRequest.class);
         cookie = new Cookie(IndexName.SESSIONID, "abcd");
     }
 
-    //store the log into the database when index with name same as the
-    //file name does not exist in the database and then trying to
-    //store another file with same fileName
+    //append sessionID to fileName to get indexName
     @Test
-    public void checkAndStoreLogInAlreadyExistingFileTest()
-    throws IOException {
-        String log = "error2";
-        String fileName = "samplefile1";
-        when(request.getCookies()).thenReturn(new Cookie[] {cookie});
-        String expected = String.format(
-            storeLogs.FILE_STORED_TEMPLATE_RESPONSE, fileName);
-        String actual = storeLogs.checkAndStoreLog(request, fileName, log);
-        assertEquals(expected, actual);
-        expected = String.format(
-            storeLogs.FILE_STORED_TEMPLATE_RESPONSE, fileName + "(1)");
-        actual = storeLogs.checkAndStoreLog(request, fileName, log);
-        assertEquals(expected, actual);
+    public void getIndexNameTest() {   
+        when(request.getCookies()).thenReturn(new Cookie[]{cookie});
+        String fileName = "file1";
+        String expected = "abcdfile1";
+        String actual = IndexName.getIndexName(request,fileName);
+        Assert.assertEquals(expected, actual);
     }
 
+    //remove sessionID from indexName to get fileName
+    @Test
+    public void getFileNameTest() { 
+        when(request.getCookies()).thenReturn(new Cookie[]{cookie});
+        String indexName = "abcdfile1";
+        String expected = "file1";
+        String actual = IndexName.getFileName(request,indexName);
+        Assert.assertEquals(expected, actual);
+    } 
 }
